feat(init): at no point are the user temp files missing...
Previously initialization started with removing the regular and global
user abbreviations temp files. They were recreated with fresh data at the end
of initialization. In the interim, other sessions would not be able to source
the temp files.

Now rather than delete and recreate the files, they are overwritten at the end
of initialization.