Respect multiline Git aliases
It is possible to define a Git alias that contains line breaks.
Previously, this would result in zsh-abbr attempting to parse each
individual line as if it were a separate alias, resulting in lots of
errors. The cause of this was splitting the output of git config
--get-regexp on line breaks.

This change splits the output of git config --get-regexp on newlines
followed by the string "alias.", thereby ensuring that an alias that
takes multiple lines gets treated as a unit.

Unfortunately this splitting doesn't get rid of the prefix "alias." on
the very first alias, and so we need to retain the substitution that
gets rid of this prefix.