Support inline expansion of Git abbreviations
The existing Git alias functionality generates global abbreviations
prefixed with "g". This is not ideal as it can clobber many other
commands or aliases.

This commit introduces functionality that allows for inline expansion of
Git abbreviations, provided they are typed on a line that starts with
the word "git". Suppose you have a Git alias fp="fetch --prune"; typing
"git fp" and pressing Space will expand to "git fetch --prune". However,
it will not expand if the first word of the command line is not "git".

It introduces a --git flag, analogous to the --global flag, that can be
used to add a Git abbreviation.